+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog of lizard-auth-server
 2.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added ``/api/v2/check_credentials/`` for checking credentials, analogous
+  to the old ``/api/authenticate/``. Used by APIs to simply verify
+  username/password credentials without any html interaction and redirects.
 
 
 2.1 (2016-08-30)